extract action buttons component
