+++ apps/thundermoon_web/lib/thundermoon_web/live/game_of_life_live/action_buttons_component.html.leex
@@ -1,0 +1,14 @@
+<div class="action-buttons">
+  <%= start_button(@socket, "start-game-of-life-button", @started) %>
+  <a href="#" id="recreate-button" phx-click="recreate" phx-target="<%= @myself %>" class="button button-icon button-warning">
+    <i class="la la-2x la-reply"></i> reset
+  </a>
+  <a href="#" id="clear-button" phx-click="clear" phx-target="<%= @myself %>" class="button button-icon button-warning">
+    <i class="la la-2x la-eraser"></i> clear
+  </a>
+  <%= if can?(@current_user, :create, GameOfLife) do %>
+    <a href="#" id="restart-button" phx-click="restart" phx-target="<%= @myself %>" class="button button-icon button-warning">
+      <i class="la la-2x la-reply-all"></i> recreate
+    </a>
+  <% end %>
+</div>
